% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rmapspam.R
\name{build_rmapspam}
\alias{build_rmapspam}
\title{Create rspam folder structure and main scripts.}
\usage{
build_rmapspam(path = NULL)
}
\value{
creates main mapspam folders and copies scripts into them
}
\description{
Create rspam folder structure and main scripts.
}
\examples{
xbuild_rmapspam

}
