% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stat.R
\name{check_stat}
\alias{check_stat}
\title{Checks consistency of subnational statistics}
\usage{
check_stat(df, param, out = F)
}
\arguments{
\item{df}{tbl or data.frame}

\item{param}{Object of type \code{spam_par} that bundles all SPAM parameters,
including core model folders, alpha-3 country code, year, spatial
resolution, most detailed level at which subnational statistics are
available, administrative unit level at which the model is solved, type of
model and coordinate reference system.}

\item{logical}{}
}
\value{
data.frame \code{df}.
}
\description{
Subnational statistics must be of class \code{subnat}
}
\examples{

}
