% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{combine_results}
\alias{combine_results}
\title{Harvest results created after running SPAMc with GAMS}
\usage{
combine_results(param)
}
\description{
Harvest results created after running SPAMc with GAMS
}
