% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_cl.R
\name{harmonize_cl}
\alias{harmonize_cl}
\title{Function that iterates over adm level starting with the most detailed and
update cl so it is in line with pa.}
\usage{
harmonize_cl(df, adm_code, param)
}
\description{
Function that iterates over adm level starting with the most detailed and
update cl so it is in line with pa.
}
