% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_results.R
\name{harvest_results}
\alias{harvest_results}
\title{Harvest results created after running SPAMc with GAMS}
\usage{
harvest_results(param)
}
\description{
Harvest results created after running SPAMc with GAMS
}
