% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_intermediate_data.R
\name{load_intermediate_data}
\alias{load_intermediate_data}
\title{Load intermediate output in line with solve level for further processing}
\usage{
load_intermediate_data(fl, adm_code, param, local = TRUE, mess = TRUE)
}
\description{
Load intermediate output in line with solve level for further processing
}
