% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapspam2globiom.R
\docType{package}
\name{mapspam2globiom-package}
\alias{mapspam2globiom}
\alias{mapspam2globiom-package}
\title{mapspam2globiom: An R package to create crop distribution maps as input for GLOBIOM}
\description{
The aim of the mapspam2globiom R package is to facilitate the creation of country level crop distribution maps, which can be used as input by the IIASA's Global Biosphere Management Model (GLOBIOM) for a selected country. mapspam2globiom includes several functions to aggregate crop distribution maps that were created with the mapspamc package (Van Dijk et al. 2022) to the GLOBIOM input format. mapspamc was specifically developed to create national crop distribution maps using the Spatial Production Allocation Model (SPAM) (You and Wood 2006; You, Wood, and Wood-Sichra 2009; You et al. 2014; Yu et al. 2020).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iiasa/mapspam2globiom}
  \item Report bugs at \url{https://github.com/iiasa/mapspam2globiom/issues}
}

}
\author{
\strong{Maintainer}: Michiel Van Dijk \email{michiel.vandijk@wur.nl} (\href{https://orcid.org/0000-0002-5207-7304}{ORCID})

Other contributors:
\itemize{
  \item Esther Boere \email{boere@iiasa.ac.at} (\href{https://orcid.org/0000-0003-3863-8731}{ORCID}) [contributor]
  \item Amanda Palazzo \email{palazzo@iiasa.ac.at} (\href{https://orcid.org/0000-0001-8167-9403}{ORCID}) [contributor]
}

}
\keyword{internal}
