% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bs_py.R
\name{prepare_bs_yg}
\alias{prepare_bs_yg}
\title{Prepare biophysical suitability and potential yield input}
\usage{
prepare_bs_yg(var, param)
}
\description{
Prepare biophysical suitability and potential yield input
}
