% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cropland.R
\name{prepare_cropland}
\alias{prepare_cropland}
\title{Prepares the synergy cropland area for SPAM}
\usage{
prepare_cropland(param)
}
\arguments{
\item{param}{list that bundles SPAM parameters, including core model folders,
alpha-3 country code, year, spatial resolution, most detailed level at which
subnational statistics are available, administrative unit level at which the
model is solved, type of model and coordinate reference system.}
}
\description{
Prepares the synergy cropland area for SPAM
}
\examples{

}
