% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pa_stat.R
\name{prepare_pa_stat}
\alias{prepare_pa_stat}
\title{Prepares the physical area subnational statistics so they can be used as input
for SPAM.}
\usage{
prepare_pa_stat(adm_code, ha, fs, ci, param)
}
\arguments{
\item{adm_code}{vector with administrative unit codes. This is either a single value
equal to the country alpha-3 country code if model solve level is set to 0
or a vector with all level 1 administrative unit codes if the model solve
level is set to 1.}

\item{ha}{tbl or data.frame with harvest area statistics}

\item{fs}{tbl or data.frame with farming system shares}

\item{ci}{tbl or data.frame with cropping intensity statistics}

\item{param}{list that bundles SPAM parameters, including core model
folders, alpha-3 country code, year, spatial resolution, most detailed level
at which subnational statistics are available, administrative unit level at
which the model is solved, type of model and coordinate reference system.}
}
\description{
To estimate the physical crop area for each farming system harvest area (ha)
statistics are combined with information on farmings sytem shares (fs) and
cropping intensity (ci). Depending on how the model is solved, the physical
area statistics are saved at the administrative unit level 0/country level
(model solve level 0) or at the level 1 administrative unit level (model solve
level 1).
}
\details{
\code{prepare_ha_stat} combines ha, fs and ci statistics and saves two files in csv
format: (1) physical area (pa) and (2) physical area broken down by farming
systems (pa_fs). Results are saved in the subfolders that are located in the
the \code{processed_data/intermediate} folder.
}
\examples{

}
