% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_score.R
\name{prepare_score}
\alias{prepare_score}
\title{Prepares score for all farming systems, crops and grid cells}
\usage{
prepare_score(param)
}
\description{
Prepares score for all farming systems, crops and grid cells
}
