% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scores.R
\name{prepare_scores}
\alias{prepare_scores}
\title{Prepares score for all farming systems, crops and grid cells}
\usage{
prepare_scores(param)
}
\description{
Prepares score for all farming systems, crops and grid cells
}
