% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_stat.R
\name{rebalance_adm}
\alias{rebalance_adm}
\title{Rebalances subnational statistics so they are consistent}
\usage{
rebalance_adm(df, spam_par)
}
\arguments{
\item{tbl}{or data.frame}

\item{par}{}
}
\value{
same class as \code{df}.
}
\description{
Rebalances subnational statistics so they are consistent
}
\examples{

}
