% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_stat.R
\name{rebalance_stat}
\alias{rebalance_stat}
\title{Rebalances subnational statistics so they are consistent}
\usage{
rebalance_stat(df, param)
}
\arguments{
\item{df}{tbl or data.frame}

\item{param}{list that bundles SPAM parameters, including core model
folders, alpha-3 country code, year, spatial resolution, most detailed level
at which subnational statistics are available, administrative unit level at
which the model is solved, type of model and coordinate reference system.}
}
\value{
same class as \code{df}.
}
\description{
Rebalances subnational statistics so they are consistent
}
\examples{

}
