% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_spam_par.R
\name{set_spam_par}
\alias{set_spam_par}
\title{Sets spam parameters}
\usage{
set_spam_par(
  spam_path = NULL,
  raw_path = NULL,
  iso3c_sel = NULL,
  year_sel = NULL,
  grid_sel = "5min",
  adm_sel = 1,
  solve_sel = 0,
  model_sel = "max_score"
)
}
\arguments{
\item{spam_path}{character string with the main SPAM folder. Note that R uses
forward slash or double backslash to separate folder names.}

\item{raw_path}{character string with the raw data folder. This makes it
possible to store the raw data on a server. If \code{raw_path} is not specified
it is automatically set to the default raw data folder in the main model
folder.}

\item{iso3c_sel}{character string with the three letter ISO 3166-1 alpha-3
country code, also referred to as iso3c. A list of country codes can be
found in \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{Wikipedia}.}

\item{year_sel}{numeric with the reference year for SPAM.}

\item{grid_sel}{character with the resolution of SPAM. Accepted inputs are
"5min" (default) and "30sec".}

\item{adm_sel}{integer with the level up to which subnational statistics are
available. Accepted inputs are 0 (only national level data), 1 (national
level and first level administrative unit - default) and 2 (national level,
first and second level administrative unit).}

\item{solve_sel}{integer that indicates the level at which the model is
solved. Accepted inputs are 0 (model is run at national level - default) and
1 (model is solved for each first level administrative unit separately).
level and first level administrative unit - default)}

\item{model_sel}{character that specifies the type of model that is run.
Accepted inputs are "max_score" and "min_entropy". See package documentation
for more information.}
}
\value{
spam_par object
}
\description{
\code{set_spam_par} sets all required parameters for spam to run, including: model
folders, country, year, resolution, administrative unit information, solve
level and model type.
}
\details{
\code{set_spam_par} creates an object of class \code{spam_par}, which bundles all
required spam parameters set by the user, including: core model folders,
country, year, resolution, depth of administrative units for which subnational
statistics are available, level at which the model is solved and type of
model.

\code{\link{create_spam_folders}} needs to be run before as it creates all
core folders, including the \code{parameters} folder where \code{set_spam_par} output is
stored. An error will be thrown if these essential folders do not exist when
the function is run.

\code{\link[countrycode]{countrycode}} is used to determine the full country
name, three digit country code, three digit FAO country code and continent on
the basis of the alpha-3 country code. This information is required to extract
country specific information from several datasets.
}
\examples{
\dontrun{
set_spam_par(spam_path = "C:/Users/dijk158/spam_mwi",
 iso3c_sel = "MWI", year_sel = 2010, grid_sel = "5min", adm_sel = 1,
 solve_sel = 0, model_sel = "max_score")
}

}
