% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_harmonized_inputs.R
\name{split_harmonized_inputs}
\alias{split_harmonized_inputs}
\title{Harmonizes data in line with solve level}
\usage{
split_harmonized_inputs(adm_code, param)
}
\description{
Harmonizes data in line with solve level
}
