% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_stat.R
\name{split_stat}
\alias{split_stat}
\title{Function to combine ha, fs and ci and split
Need to split first and then combine as in split version, adm specific fs and ci are used}
\usage{
split_stat(adm_code, ha, fs, ci, param)
}
\description{
Function to combine ha, fs and ci and split
Need to split first and then combine as in split version, adm specific fs and ci are used
}
