% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cl.R
\name{update_cl}
\alias{update_cl}
\title{Function that iterates over adm level starting with the most detailed and
update cl so it is in line with pa.}
\usage{
update_cl(df, param)
}
\description{
Function that iterates over adm level starting with the most detailed and
update cl so it is in line with pa.
}
