% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_spam_par.R
\name{validate_spam_par}
\alias{validate_spam_par}
\title{Validates spam_par class}
\usage{
validate_spam_par(spam_par)
}
\arguments{
\item{spam_par}{list that bundles SPAM parameters, including core model
folders, alpha-3 country code, year, spatial resolution, most detailed level
at which subnational statistics are available, administrative unit level at
which the model is solved, type of model and coordinate reference system.}
}
\description{
Validates spam_par class
}
\examples{
\dontrun{}
}
