% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_4panel.R
\name{view_4panel}
\alias{view_4panel}
\title{Function to compare crop maps per system in a panel}
\usage{
view_4panel(crp, vr, sy = c("S", "L", "H", "I"), param)
}
\description{
Function to compare crop maps per system in a panel
}
