% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_stack.R
\name{view_stack}
\alias{view_stack}
\title{Compare all farming systems of a selected crop in the browser using stack}
\usage{
view_stack(crp, vr, sy = c("S", "L", "H", "I"))
}
\description{
Compare all farming systems of a selected crop in the browser using stack
}
